fix: address PR #3219 review comments
- Extract tag-rule validators into apps/evaluations/rule_validation.py so
  models.py can import them at module level (removes the local-import
  workaround for the tagging ↔ models cycle).
- Give ConditionType a coerce_value staticmethod and matches instance
  method so form/validation callers stop duplicating type-coercion and
  predicate logic; tagging.matches() now delegates.
- EvaluatorTagRuleForm: lift all local imports to the module level,
  initialise _resolved_tag_name in __init__, and split the long clean()
  into _clean_required_row_fields, _clean_condition_value, and
  _validate_against_schema helpers.
- apply_rules_to_result no longer removes tags — tag removal on rerun
  is deferred to a follow-up PR. Drop the two removal-semantics tests.
